+++ app/Http/Controllers/CategoriasController.php
@@ -30,17 +30,28 @@ class CategoriasController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Categoria $categoria)
+    public function show($id_categoria)
     {
-        $categoria = Categoria::find($id_categoria);
+        $categoria = Categoria::where('id_categoria', $id_categoria)->first();
+
+        if (!$categoria) {
+            return response()->json(['error' => 'Categoria no encontrada'], 404);
+        }
+
         return $categoria;
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Categoria $categoria)
+    public function update(Request $request, $categoria)
     {
+        $categoria = Categoria::where('id_categoria', $id_categoria)->first();
+
+        if (!$categoria) {
+            return response()->json(['error' => 'Categoria no encontrada'], 404);
+        }
+
         $categoria->nombre_categoria = $request->nombre_categoria;
         $categoria->color = $request->color;
         $categoria->save();
@@ -50,8 +61,13 @@ class CategoriasController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Categoria $categoria)
+    public function destroy($id_categoria)
     {
+        $categoria = Categoria::where('id_categoria',$id_categoria);
+        if (!$categoria) {
+            return response()->json(['error' => 'Categoria no encontrada.'], 404);
+        }
+
         return $categoria->delete();
     }
 }
